完成评论功能

+++ src/main/java/ml/dengshen/community/community/controller/PublishController.java
@@ -28,7 +28,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable("id") Integer id,
+    public String edit(@PathVariable("id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -42,7 +42,7 @@ public class PublishController {
     public String doPublish(@RequestParam("title") String title,
                             @RequestParam("description") String description,
                             @RequestParam("tag") String tag,
-                            @RequestParam("id") Integer  id,
+                            @RequestParam("id") Long  id,
                             HttpServletRequest request,
                             Model model) {
         model.addAttribute("title", title);
